+++ src/lib/components/form/components/field/field.component.ts
@@ -12,6 +12,7 @@ import { FormService } from '../../services/form.service';
 import { FormCheckboxFieldComponent } from './field-checkbox/field-checkbox.component';
 import { FormChipsFieldComponent } from './field-chips/field-chips.component';
 import { FormDateFieldComponent } from './field-date/field-date.component';
+import { FormFileFieldComponent } from './field-file/field-file.component';
 import { FormInfoFieldComponent } from './field-info/field-info.component';
 import { FormNumberFieldComponent } from './field-number/field-number.component';
 import { FormPasswordFieldComponent } from './field-password/field-password.component';
@@ -28,6 +29,7 @@ import { FormTextareaFieldComponent } from './field-textarea/field-textarea.comp
         FormCheckboxFieldComponent,
         FormChipsFieldComponent,
         FormDateFieldComponent,
+        FormFileFieldComponent,
         FormInfoFieldComponent,
         FormNumberFieldComponent,
         FormPasswordFieldComponent,